sections/02_reflection_agent/projects/pyproject.toml
[project]
name = "projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "langchain-ollama>=1.0.0",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.2.1",
]
